requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pdfrw==0.4
Pillow==7.2.0
pylint==2.6.0
python-bidi==0.4.2
python-dotenv==0.14.0
reportlab==3.5.51
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
